+++ bin/alex.dart
@@ -1,5 +1,7 @@
+import 'dart:io';
+
 import 'package:alex/app.dart';
 
-Future<int> main(List<String> args) {
-  return run(args);
+Future<void> main(List<String> args) async {
+  exitCode = await run(args);
 }